refactor: modularize native facade wrappers
